+++ __tests__/presentation-3-parser/cookbook.tests.ts
@@ -3,7 +3,7 @@ import { promises } from 'node:fs';
 import { cwd } from 'node:process';
 import { join } from 'path';
 const { readFile } = promises;
-import { normalize, serialize, serializeConfigPresentation3 } from '../../src/presentation-3';
+import { normalize, serialize, serializeConfigPresentation3 } from '../../src';
 
 const prWaitingForMerge: string[] = [
   // '0219-using-caption-file', // https://github.com/IIIF/cookbook-recipes/pull/340