+++ src/presentation-3/normalize.ts
@@ -1,4 +1,4 @@
-import { Traverse } from './traverse';
+import { TraversalContext, Traverse } from './traverse';
 import {
   Annotation,
   AnnotationPage,
@@ -14,7 +14,6 @@ import {
   Service,
 } from '@iiif/presentation-3';
 import {
-  EMPTY,
   emptyAgent,
   emptyAnnotationPage,
   emptyCanvas,
@@ -35,6 +34,7 @@ import {
   ResourceProviderNormalized,
 } from '@iiif/presentation-3-normalized';
 import { isSpecificResource } from '../shared/is-specific-resource';
+import { EMPTY, HAS_PART, PART_OF, WILDCARD } from './utilities';
 
 export const defaultEntities = {
   Collection: {},
@@ -76,15 +76,21 @@ function getResource(entityOrString: PolyEntity, type: string): Reference {
   return entityOrString as Reference;
 }
 
-function mapToEntities(entities: Record<string, Record<string, NormalizedEntity>>) {
+function mapToEntities(entities: Record<string, Record<string, NormalizedEntity>>, topLevel: any) {
   return <T extends Reference | string>(type: string, defaultStringType?: string) => {
     const storeType = entities[type] ? entities[type] : {};
-    return (r: T): T => {
+    return (r: T, context: TraversalContext): T => {
       const resource = getResource(r, defaultStringType || type);
       if (resource && resource.id && type) {
         storeType[resource.id] = storeType[resource.id]
-          ? (mergeEntities(storeType[resource.id], resource) as any)
-          : Object.assign({}, resource);
+          ? (mergeEntities(storeType[resource.id], resource, {
+              parent: context.parent,
+              isTopLevel: topLevel.id === resource.id,
+            }) as any)
+          : mergeEntities({ id: resource.id, type: resource.type } as any, resource, {
+              parent: context.parent,
+              isTopLevel: topLevel.id === resource.id,
+            });
         return {
           id: resource.id,
           type: type === 'ContentResource' ? type : resource.type,
@@ -95,7 +101,7 @@ function mapToEntities(entities: Record<string, Record<string, NormalizedEntity>
   };
 }
 
-function merge(existing: any, incoming: any): any {
+export function merge(existing: any, incoming: any, context?: { parent?: any; isTopLevel?: boolean }): any {
   if (!incoming) {
     // Falsy values are ignored
     return existing;
@@ -133,14 +139,92 @@ function merge(existing: any, incoming: any): any {
     // For objects, we check the existing object for non-existing or "empty"
     // properties and use the value from the incoming object for them
     const merged = { ...existing };
+    const added: string[] = [];
+    const unchanged: string[] = [];
+    const existingKeys = Object.keys(existing).filter((key) => key !== HAS_PART && key !== 'id' && key !== 'type');
+    const previouslyChangedValues: any = {};
+    const incomingChangedValues: any = {};
     for (const [key, val] of Object.entries(incoming)) {
+      if (key === HAS_PART || key === 'id' || key === 'type') {
+        continue;
+      }
       const currentVal = merged[key];
-      if (currentVal === EMPTY || !currentVal) {
+      if (currentVal === val) {
+        unchanged.push(key);
+      } else if (currentVal === EMPTY || !currentVal) {
+        added.push(key);
         merged[key] = val;
       } else {
+        if (currentVal && val) {
+          previouslyChangedValues[key] = currentVal;
+          incomingChangedValues[key] = val;
+        }
         merged[key] = merge(currentVal, val);
+        if (merged[key] === previouslyChangedValues[key]) {
+          unchanged.push(key);
+          delete previouslyChangedValues[key];
+        }
       }
     }
+
+    if (context && ((context.parent && context.parent.id) || context.isTopLevel)) {
+      const newHasPart: any[] = [];
+      const part: any = {};
+      if (context.parent) {
+        part[PART_OF] = context.parent.id;
+      } else if (context.isTopLevel) {
+        part[PART_OF] = existing.id;
+      }
+
+      if (merged[HAS_PART] && merged[HAS_PART].length) {
+        const noExplicit = !merged[HAS_PART].find((r: any) => r['@explicit']);
+        const hasDiverged = added.length > 0 || unchanged.length !== existingKeys.length;
+        // We already have one, it may conflict here.
+        // 1. Fix the first part.
+        if (noExplicit && hasDiverged) {
+          for (const item of merged[HAS_PART]) {
+            const first = { ...item };
+            const changedKeys = Object.keys(previouslyChangedValues);
+            if (first) {
+              first['@explicit'] = true;
+              for (const addedProperty of existingKeys) {
+                if (addedProperty !== HAS_PART) {
+                  first[addedProperty] = WILDCARD;
+                }
+              }
+              for (const changedKey of changedKeys) {
+                first[changedKey] = previouslyChangedValues[changedKey];
+              }
+            }
+            newHasPart.push(first);
+          }
+        } else {
+          newHasPart.push(...merged[HAS_PART]);
+        }
+
+        if (hasDiverged) {
+          // Add the framing.
+          const changedKeys = Object.keys(incomingChangedValues);
+          part['@explicit'] = true;
+          for (const addedProperty of added) {
+            part[addedProperty] = WILDCARD;
+          }
+          for (const unchangedValue of unchanged) {
+            part[unchangedValue] = WILDCARD;
+          }
+          for (const changedKey of changedKeys) {
+            part[changedKey] = incomingChangedValues[changedKey];
+          }
+        }
+      }
+
+      part.id = merged.id;
+      part.type = merged.type;
+      newHasPart.push(part);
+
+      merged[HAS_PART] = newHasPart;
+    }
+
     return merged;
   } else if (existing) {
     return existing;
@@ -148,14 +232,29 @@ function merge(existing: any, incoming: any): any {
   return incoming;
 }
 
-export function mergeEntities(existing: NormalizedEntity, incoming: any): NormalizedEntity {
+export function mergeEntities(
+  existing: NormalizedEntity,
+  incoming: any,
+  context?: { parent?: any; isTopLevel?: boolean }
+): NormalizedEntity {
   if (typeof existing === 'string') {
     return existing;
   }
   if (incoming.id !== (existing as any).id || incoming.type !== (existing as any).type) {
-    throw new Error('Can only merge entities with identical identifiers and type!');
+    if (incoming.type === 'ImageService3') {
+      return incoming;
+    }
+    if ((existing as any).type === 'ImageService3') {
+      return existing;
+    }
+
+    throw new Error(
+      `Can only merge entities with identical identifiers and type! ${incoming.type}(${incoming.id}) => ${
+        (existing as any).type
+      }(${(existing as any).id})`
+    );
   }
-  return merge({ ...existing }, incoming);
+  return merge({ ...existing }, incoming, context);
 }
 
 function recordTypeInMapping(mapping: Record<string, string>) {
@@ -379,7 +478,7 @@ export function normalize(unknownEntity: unknown) {
   const entity = convertPresentation2(unknownEntity);
   const entities = getDefaultEntities();
   const mapping = {};
-  const addToEntities = mapToEntities(entities);
+  const addToEntities = mapToEntities(entities, entity);
   const addToMapping = recordTypeInMapping(mapping);
 
   const traversal = new Traverse({